refactor(identity): the field-mask loader lives in its own file
