+++ backend/internal/modules/identity/service.go
@@ -464,31 +464,6 @@ func loadGrants(ctx context.Context, tx pgx.Tx, userID ids.UserID) (roles []stri
 	return roles, teams, perms, err
 }
 
-// loadFieldMasks reads the columns the principal's roles withhold — the
-// union over every role held, additive like the grants: a role that masks a
-// field masks it for the seat, whatever another role says.
-func loadFieldMasks(ctx context.Context, tx pgx.Tx, roleKeys []string) ([]principal.FieldMask, error) {
-	if len(roleKeys) == 0 {
-		return nil, nil
-	}
-	rows, err := tx.Query(ctx,
-		`SELECT DISTINCT object, field, condition FROM field_mask WHERE role_key = ANY($1) ORDER BY object, field, condition`,
-		roleKeys)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var masks []principal.FieldMask
-	for rows.Next() {
-		var m principal.FieldMask
-		if err := rows.Scan(&m.Object, &m.Field, &m.Condition); err != nil {
-			return nil, err
-		}
-		masks = append(masks, m)
-	}
-	return masks, rows.Err()
-}
-
 // rawTeamIDs widens typed team ids to the untyped []ids.UUID the kernel
 // principal and the authz port carry — the row-scope seams stay untyped
 // (they compare team membership against polymorphic scope clauses).
